third_party/WebKit/LayoutTests/http/tests/websocket/workers/close-before-open.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script type="text/javascript">

var test = async_test("Test that an error is shown on the console when close() is called on a WebSocket synchronously right after it's constructed.");
test.step(function()
{
    var worker = new Worker("resources/close-before-open.js");
    worker.onmessage = test.step_func(function(evt)
    {
        assert_equals(evt.data, "DONE", "evt.data");
        test.done();
    });
});

</script>
</head>
<body>
</body>
</html>
