third_party/WebKit/LayoutTests/http/tests/websocket/workers/close-before-open-expected.txt

CONSOLE WARNING: line 18: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: WebSocket is closed before the connection is established.
This is a testharness.js-based test.
PASS Test that an error is shown on the console when close() is called on a WebSocket synchronously right after it's constructed. 
Harness: the test ran to completion.
